\name{alkfos}
\alias{alkfos}
\docType{data}
\title{Alkaline phosphatase data}
\description{
  Repeated measurements of alkaline phosphatase in a randomized trial of
  Tamoxifen treatment of breast cancer patients. 
}
\usage{alkfos}
\format{
  A data frame with 43 observations on the following 8 variables.
  \describe{
    \item{\code{grp}}{a numeric vector, group code (1=placebo, 2=Tamoxifen).}
    \item{\code{c0}}{a numeric vector, concentration at baseline.}
    \item{\code{c3}}{a numeric vector, concentration after 3 months.}
    \item{\code{c6}}{a numeric vector, concentration after 6 months.}
    \item{\code{c9}}{a numeric vector, concentration after 9 months.}
    \item{\code{c12}}{a numeric vector, concentration after 12 months.}
    \item{\code{c18}}{a numeric vector, concentration after 18 months.}
    \item{\code{c24}}{a numeric vector, concentration after 24 months.}
  }
}
\source{
  Original data.
}
\references{
 B. Kristensen et al. (1994),
 Tamoxifen and bone metabolism in postmenopausal low-risk breast
 cancer patients: a randomized study. \emph{Journal of Clinical Oncology}, 12(2):992--997.
}
\keyword{datasets}
