\name{IgM}
\alias{IgM}
\title{Immunoglobulin G}
\description{
  Serum IgM
  in 298 children aged 6 months to 6 years.
}
\usage{IgM}
\format{A single numeric vector (g/l).}
\source{
  D.G. Altman (1991), \emph{Practical Statistics for Medical Research}, Table
  3.2, Chapman & Hall.
}
\examples{
stripchart(IgM,method="stack")
}
\keyword{datasets}
