\name{zelazo}
\alias{zelazo}
\title{Age at walking}
\description{The \code{zelazo} object is a list with four components.} 
\usage{zelazo}
\format{
  This is a list containing data on age at
  walking (in months) for four groups of infants: 
\describe{
\item{\code{active}}{test group receiving active training; these children had
  their walking and placing reflexes
  trained during four three-minute sessions that took place every day
  from their second to their eighth week of life.}
\item{\code{passive}}{passive training group; these children received the same types of
  social and gross motor stimulation, 
  but did not have their specific walking and placing reflexes
  trained. }
\item{\code{none}}{no training; these children  had no special training, but
  were tested along with the children who underwent active or passive
  training.}
\item{\code{ctr.8w}}{eighth-week controls; these children had no training and were only
  tested at the age of 8 weeks. }
}
}
\source{P.R. Zelazo, N.A. Zelazo, and S. Kolb (1972),
  ``Walking'' in the newborn, \emph{Science}, 176: 314--315.
}
\note{When asked to enter these data from a text source, many
students will use one vector per group and will need to reformat data
into a data frame for some uses. The rather unusual format of this data set
mimics that situation. 
}
\keyword{datasets}
