\name{lung}
\alias{lung}
\title{Methods for determining lung volume}
\description{
The \code{lung} data frame has 18 rows and 3 columns. It contains data
on three different  methods  of determining human
  lung volume.
}
\usage{lung}
\format{
  This data frame contains the following columns:
  \describe{
    \item{\code{volume}}{a numeric vector, measured lung volume.}
    \item{\code{method}}{a factor with levels \code{A}, \code{B}, and \code{C}. }
    \item{\code{subject}}{a factor with levels \code{1}--\code{6}.}
  }
}
\source{
 \emph{Anon. (1977),  Exercises in Applied Statistics}, Exercise 4.15, Dept.\ of
 Theoretical Statistics,  
 Aarhus University.
}
\keyword{datasets}
